+++ cpp/src/arrow/CMakeLists.txt
@@ -750,7 +750,6 @@ if(ARROW_COMPUTE)
        compute/kernels/aggregate_quantile.cc
        compute/kernels/aggregate_tdigest.cc
        compute/kernels/aggregate_var_std.cc
-       compute/kernels/codegen_internal.cc
        compute/kernels/hash_aggregate.cc
        compute/kernels/hash_aggregate_numeric.cc
        compute/kernels/hash_aggregate_pivot.cc
