+++ cpp/src/arrow/compute/kernels/chunked_internal.h
@@ -25,7 +25,7 @@
 
 #include "arrow/array.h"
 #include "arrow/chunk_resolver.h"
-#include "arrow/compute/kernels/codegen_internal.h"
+#include "arrow/compute/codegen_internal.h"
 #include "arrow/util/span.h"
 #include "arrow/util/visibility.h"
