+++ cpp/src/arrow/compute/kernels/common_internal.h
@@ -31,9 +31,9 @@
 #include "arrow/buffer.h"
 #include "arrow/chunked_array.h"
 
+#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/function.h"
 #include "arrow/compute/kernel.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/registry.h"
 #include "arrow/datum.h"
 #include "arrow/memory_pool.h"
