+++ cpp/src/arrow/compute/kernels/pivot_internal.cc
@@ -24,8 +24,8 @@
 #include "arrow/array/array_primitive.h"
 #include "arrow/array/builder_binary.h"
 #include "arrow/compute/cast.h"
+#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/exec.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/row/grouper.h"
 #include "arrow/scalar.h"
 #include "arrow/type_traits.h"
