+++ cpp/src/arrow/compute/kernels/scalar_arithmetic.cc
@@ -27,7 +27,6 @@
 #include "arrow/compute/cast.h"
 #include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/kernels/base_arithmetic_internal.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/common_internal.h"
 #include "arrow/compute/kernels/util_internal.h"
 #include "arrow/type.h"
