+++ cpp/src/arrow/compute/kernels/scalar_round.cc
@@ -28,7 +28,6 @@
 #include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/kernel.h"
 #include "arrow/compute/kernels/base_arithmetic_internal.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/common_internal.h"
 #include "arrow/compute/kernels/util_internal.h"
 #include "arrow/type.h"
