+++ cpp/src/arrow/compute/kernels/scalar_string_test.cc
@@ -24,8 +24,8 @@
 #include <gtest/gtest.h>
 
 #include "arrow/compute/api_scalar.h"
+#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/exec.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/test_util_internal.h"
 #include "arrow/testing/gtest_util.h"
 #include "arrow/type.h"
