+++ cpp/src/arrow/compute/kernels/vector_pairwise.cc
@@ -22,11 +22,11 @@
 
 #include "arrow/array/builder_base.h"
 #include "arrow/compute/api_vector.h"
+#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/exec.h"
 #include "arrow/compute/function.h"
 #include "arrow/compute/kernel.h"
 #include "arrow/compute/kernels/base_arithmetic_internal.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/registry.h"
 #include "arrow/compute/util.h"
 #include "arrow/status.h"
