+++ cpp/src/arrow/compute/kernels/vector_selection_internal.cc
@@ -30,7 +30,6 @@
 #include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/function.h"
 #include "arrow/compute/kernel.h"
-#include "arrow/compute/kernels/codegen_internal.h"
 #include "arrow/compute/kernels/vector_selection_internal.h"
 #include "arrow/compute/registry.h"
 #include "arrow/type.h"
