+++ cpp/src/arrow/compute/kernels/vector_selection_take_internal.cc
@@ -27,7 +27,7 @@
 #include "arrow/buffer_builder.h"
 #include "arrow/chunked_array.h"
 #include "arrow/compute/api_vector.h"
-#include "arrow/compute/kernels/codegen_internal.h"
+#include "arrow/compute/codegen_internal.h"
 #include "arrow/compute/kernels/gather_internal.h"
 #include "arrow/compute/kernels/vector_selection_internal.h"
 #include "arrow/compute/kernels/vector_selection_take_internal.h"
